% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_wiche.R
\name{lookup_wiche}
\alias{lookup_wiche}
\title{Lookup WICHE Projections}
\usage{
lookup_wiche(location = "us", sector = "public", race = "all",
  gender = "all", grade = "graduates", year = NULL)
}
\arguments{
\item{location}{state abbreviation in lowercase or 'us' for the entire united states}

\item{sector}{identifies all, public or non-public schools}

\item{race}{identifies all or particular races/ethnicity}

\item{grade}{grade level number or 'graduates'}

\item{year}{academic year ending, so 2015-2016 would be expresses as 2016}

\item{n}{the estimate of the number of students enrolled}
}
\value{
A data frame with:
}
\description{
Quick lookup function for the WICHE table
}
\examples{

lookup_wiche(year = 2012:2016)

  location sector  race     grade year n
1       us public  all          8 2012 3053230
2       us public  all          9 2013 3022879
3       us public  all         10 2014 3022879
4       us public  all         11 2015 2973914
5       us public  all         12 2016 3000793
}
\keyword{WICHE,}
\keyword{cohort}
\keyword{high}
\keyword{projections,}
\keyword{school}

