% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\docType{data}
\name{states}
\alias{states}
\title{Information about US States}
\format{A data frame with 53 rows and 6 variables:
\describe{
  \item{state_id}{state abbreviation in lowercase that can link back to student files}
  \item{abbrev}{state abbreviation in uppercase}
  \item{long}{long state name typically used in reports}
}}
\usage{
states
}
\description{
Contains reference information about US States
}
\keyword{datasets}

