% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiche.R
\docType{data}
\name{wiche}
\alias{wiche}
\title{WICHE Projections}
\format{A data frame with 193,024 rows and 6 variables:
\describe{
  \item{location}{state abbreviation in lowercase}
  \item{sector}{identifies all, public or non-public schools}
  \item{race}{identifies all or particular races}
  \item{grade}{grade level number or 'graduates'}
  \item{year}{academic year ending, so 2015-2016 would be expresses as 2016}
  \item{n}{the number of students enrolled}
}}
\source{
{http://www.wiche.edu/knocking-8th}
}
\usage{
wiche
}
\description{
High school graduate forecasts from The Western Interstate Commission for Higher Education (WICHE)
}
\examples{
head(wiche)
  location sector  race     grade year    n
1       in public asian         8 2021 2306
2       in public asian         9 2021 2407
3       in public asian        10 2021 2314
4       in public asian        11 2021 2801
5       in public asian        12 2021 2578
6       in public asian graduates 2021 2201
}
\keyword{datasets}

