% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiche_enrollments.R
\docType{data}
\name{wiche_enrollments}
\alias{wiche_enrollments}
\title{WICHE Enrollments (2016)}
\format{A data frame with 186,368 rows and 7 variables:
\describe{
  \item{location}{region, us or state abbreviation in lowercase}
  \item{sector}{identifies all, public or non-public schools}
  \item{race}{identifies all or particular races}
  \item{sector}{all schools, public schools or private schools}
  \item{grade}{grade level number or 'graduates'}
  \item{year}{academic year ending, so 2015-2016 would be expressed as 2016}
  \item{actual}{are these actual or projected enrollments}
  \item{n}{the number of students enrolled}
}}
\source{
{http://knocking.wiche.edu/data}
}
\usage{
wiche_enrollments
}
\description{
Grade forecasts from The Western Interstate Commission for Higher Education (WICHE)
}
\examples{

library(wicher)
wiche_enrollments

}
\keyword{datasets}
