% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiche_ingest_2016_enrollments.R
\name{wiche_ingest_2016_enrollments}
\alias{wiche_ingest_2016_enrollments}
\title{Ingest 2016 Enrollment Projections}
\usage{
wiche_ingest_2016_enrollments(file_location)
}
\arguments{
\item{file_location}{where the workbook is located}
}
\description{
Code reads in enrollments in the WICHE spreadsheet and puts them into
wicher format. This was used to create the `wiche_graduate_projects`
dataframe
The raw data can be located on the WICHE
Knocking at the College Door(http://knocking.wiche.edu/data)
}
